Refactorizar GlassCard, selectorSala y ConfirmacionReserva
