+++ DREAM Lab Frontend/src/components/aceternity/evervault-card.jsx
@@ -2,7 +2,7 @@ import { useMotionValue } from "framer-motion";
 import { useState, useEffect } from "react";
 import { useMotionTemplate, motion } from "framer-motion";
 import { cn } from "../../utils/cn";
-import GlassCard from "../../components/general/GlassCard";
+import GlassCard from "../../GlobalComponents/GlassCard/GlassCard";
 import MicrophoneButton from "../../Home/MicrophoneButton.jsx";
 import SendButton from "../../Home/SendButton.jsx";
 import "../../App.css";
@@ -10,93 +10,95 @@ import "../../App.css";
 import PropTypes from "prop-types";
 
 function generateRandomString(length) {
-  const characters =
-    "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  let result = "";
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * characters.length));
-  }
-  return result;
+    const characters =
+        "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+    let result = "";
+    for (let i = 0; i < length; i++) {
+        result += characters.charAt(
+            Math.floor(Math.random() * characters.length)
+        );
+    }
+    return result;
 }
 
 export const EvervaultCard = () => {
-  let mouseX = useMotionValue(0);
-  let mouseY = useMotionValue(0);
+    let mouseX = useMotionValue(0);
+    let mouseY = useMotionValue(0);
 
-  const [randomString, setRandomString] = useState("");
+    const [randomString, setRandomString] = useState("");
 
-  useEffect(() => {
-    let str = generateRandomString(8000);
-    setRandomString(str);
-  }, []);
+    useEffect(() => {
+        let str = generateRandomString(8000);
+        setRandomString(str);
+    }, []);
 
-  function onMouseMove({ currentTarget, clientX, clientY }) {
-    let { left, top } = currentTarget.getBoundingClientRect();
-    mouseX.set(clientX - left);
-    mouseY.set(clientY - top);
+    function onMouseMove({ currentTarget, clientX, clientY }) {
+        let { left, top } = currentTarget.getBoundingClientRect();
+        mouseX.set(clientX - left);
+        mouseY.set(clientY - top);
 
-    const str = generateRandomString(8000);
-    setRandomString(str);
-  }
+        const str = generateRandomString(8000);
+        setRandomString(str);
+    }
 
-  return (
-    <div
-      className={cn(
-        "p-0.5  bg-transparent aspect-square  flex items-center justify-center w-full h-full relative"
-      )}
-    >
-      <div
-        onMouseMove={onMouseMove}
-        className="group/card rounded-3xl w-full relative overflow-hidden bg-transparent flex items-center justify-center h-full"
-      >
-        <CardPattern
-          mouseX={mouseX}
-          mouseY={mouseY}
-          randomString={randomString}
-        />
-        <div className="relative z-10 flex items-center justify-center">
-          <div className="centered-container">
-            <GlassCard
-              className="inputReconocimientoVoz"
-              height="6rem"
-              width="50rem"
-              padding="0 2rem 0 2rem"
+    return (
+        <div
+            className={cn(
+                "p-0.5  bg-transparent aspect-square  flex items-center justify-center w-full h-full relative"
+            )}
+        >
+            <div
+                onMouseMove={onMouseMove}
+                className="group/card rounded-3xl w-full relative overflow-hidden bg-transparent flex items-center justify-center h-full"
             >
-              <MicrophoneButton />
-              <SendButton />
-            </GlassCard>
-          </div>
+                <CardPattern
+                    mouseX={mouseX}
+                    mouseY={mouseY}
+                    randomString={randomString}
+                />
+                <div className="relative z-10 flex items-center justify-center">
+                    <div className="centered-container">
+                        <GlassCard
+                            className="inputReconocimientoVoz"
+                            height="6rem"
+                            width="50rem"
+                            padding="0 2rem 0 2rem"
+                        >
+                            <MicrophoneButton />
+                            <SendButton />
+                        </GlassCard>
+                    </div>
+                </div>
+            </div>
         </div>
-      </div>
-    </div>
-  );
+    );
 };
 
 export function CardPattern({ mouseX, mouseY, randomString }) {
-  let maskImage = useMotionTemplate`radial-gradient(250px at ${mouseX}px ${mouseY}px, white, transparent)`;
-  let style = { maskImage, WebkitMaskImage: maskImage };
+    let maskImage = useMotionTemplate`radial-gradient(250px at ${mouseX}px ${mouseY}px, white, transparent)`;
+    let style = { maskImage, WebkitMaskImage: maskImage };
 
-  return (
-    <div className="pointer-events-none">
-      <div className="absolute inset-0 rounded-2xl  [mask-image:linear-gradient(white,transparent)] group-hover/card:opacity-50"></div>
-      <motion.div
-        className="absolute inset-0 rounded-2xl bg-gradient-to-r from-purple-500 to-blue-700 opacity-0  group-hover/card:opacity-100 backdrop-blur-xl transition duration-500"
-        style={style}
-      />
-      <motion.div
-        className="absolute inset-0 rounded-2xl opacity-0 mix-blend-overlay  group-hover/card:opacity-100"
-        style={style}
-      >
-        <p className="absolute inset-x-0 text-xs h-full break-words whitespace-pre-wrap text-white font-mono font-bold transition duration-500">
-          {randomString}
-        </p>
-      </motion.div>
-    </div>
-  );
+    return (
+        <div className="pointer-events-none">
+            <div className="absolute inset-0 rounded-2xl  [mask-image:linear-gradient(white,transparent)] group-hover/card:opacity-50"></div>
+            <motion.div
+                className="absolute inset-0 rounded-2xl bg-gradient-to-r from-purple-500 to-blue-700 opacity-0  group-hover/card:opacity-100 backdrop-blur-xl transition duration-500"
+                style={style}
+            />
+            <motion.div
+                className="absolute inset-0 rounded-2xl opacity-0 mix-blend-overlay  group-hover/card:opacity-100"
+                style={style}
+            >
+                <p className="absolute inset-x-0 text-xs h-full break-words whitespace-pre-wrap text-white font-mono font-bold transition duration-500">
+                    {randomString}
+                </p>
+            </motion.div>
+        </div>
+    );
 }
 
 CardPattern.propTypes = {
-  mouseX: PropTypes.object,
-  mouseY: PropTypes.object,
-  randomString: PropTypes.string,
+    mouseX: PropTypes.object,
+    mouseY: PropTypes.object,
+    randomString: PropTypes.string,
 };
