+++ DREAM Lab Frontend/src/pages/HomePage/components/SpeechBotCard/SpeechBotCard.jsx
@@ -7,7 +7,7 @@ import SpeechRecognition, {
 } from "react-speech-recognition";
 import TextField from "./components/TextField/TextField.jsx";
 import "./SpeechBotCard.css";
-import "../../../../components/general/GlassCard.css";
+import "../../../../GlobalComponents/GlassCard/GlassCard.css";
 
 function SpeechBotCard(props) {
     // --- Para el reconocimiento de voz ---
